refactor: CompletePage의 경로를 layout으로 변경
