+++ frontend/src/components/layout/completePage/CompletePage.styles.ts
